+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - CI-217 - Update Django Version to fix security vulnerability.
 - TT-1529 - Fix 500 error pages so links work
 - TT-1531 FAS - Logos are misaligned on trade profiles (fixed cross browser compatibility issues)
+- TT-1547 - Prevent FAS "&sectors=" getting encoded as "§ors=" by renaming to industries.
 
 ## [2019.06.05] (https://github.com/uktrade/directory-ui-supplier/releases/tag/2019.06.05)
 [Full Changelog](https://github.com/uktrade/directory-ui-supplier/compare/2019.05.28_1...2019.06.05)